[VBTB-5833][VBTB-5840] Made a simpler json response

+++ src/main/java/br/com/pokemon/interfaceadapter/gateway/PokemonListGateway.java
@@ -22,5 +22,5 @@ public interface PokemonListGateway {
     @GET
     @Path("pokemon/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    PokemonDetailResponse listPokemonById(@PathParam("id") Long id);
+    PokemonDetailResponse listPokemonById(@PathParam("id") int id);
 }
